requirements.txt
matplotlib>=3.9.4
numpy>=1.26.4
scipy>=1.13.1
pyyaml>=6.0.2